Final touches before PR to dev
